refactor: 로컬 모델 구조 개선 및 파일 재구성
- `local_model.py` 파일 삭제 및 기존 로직을 `gemma_3_4b.py`로 이동
- `local_model_generate` 참조 경로를 `src.models.gemma_3_4b`로 수정
- 불필요한 `processor` 관련 코드 제거 및 `tokenizer`로 통합
- 서버 기본 포트 수정 (9999 → 8000) (#61)
