+++ src/services/v2/description_writer.py
@@ -2,7 +2,7 @@ from typing import Tuple
 from src.schemas.v1.group_writer import GroupGenerationRequest
 # from src.core.cloud_logging import logger
 from src.core.ai_logger import get_ai_logger
-from src.services.v2.local_model import local_model_generate   #로컬 모델 호출로 교체
+from src.models.gemma_3_4b import local_model_generate   #로컬 모델 호출로 교체
 
 ai_logger = get_ai_logger()
